+++ app/src/main/java/me/rerere/rikkahub/ui/hooks/ASR.kt
@@ -9,12 +9,12 @@ import androidx.compose.ui.platform.LocalContext
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
-import me.rerere.rikkahub.data.asr.ASRController
-import me.rerere.rikkahub.data.asr.ASRProviderSetting
-import me.rerere.rikkahub.data.asr.ASRState
-import me.rerere.rikkahub.data.asr.DashScopeASRController
-import me.rerere.rikkahub.data.asr.OpenAIRealtimeASRController
-import me.rerere.rikkahub.data.asr.VolcengineASRController
+import me.rerere.asr.ASRController
+import me.rerere.asr.ASRProviderSetting
+import me.rerere.asr.ASRState
+import me.rerere.asr.providers.DashScopeASRController
+import me.rerere.asr.providers.OpenAIRealtimeASRController
+import me.rerere.asr.providers.VolcengineASRController
 import me.rerere.rikkahub.data.datastore.SettingsStore
 import me.rerere.rikkahub.data.datastore.getSelectedASRProvider
 import okhttp3.OkHttpClient
