+++ app/src/main/java/me/rerere/rikkahub/ui/pages/setting/components/ASRProviderConfigure.kt
@@ -10,7 +10,7 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
-import me.rerere.rikkahub.data.asr.ASRProviderSetting
+import me.rerere.asr.ASRProviderSetting
 import me.rerere.rikkahub.ui.components.ui.FormItem
 import me.rerere.rikkahub.ui.components.ui.OutlinedNumberInput
